Pass empty GIT_DIR env var to bower install
When bower tries to fetch a dependency via git, internally it runs a git
command. If there is a .git file or dir in the current directory, git
automatically tries to discover which repo it is in.

When mounting code inside the container, especially in the case of
mounting code which is in a git submodule, then there's a .git file that
points to the .git dir of the parent repo, which isn't mounted inside
the container. In that case, when git automatically tries to follow the
link of the .git file, it will fail since it cannot find it. By passing
this env var, this stops happening.